refactor(renderer): extract useTheme hook from App.tsx
